Add support for PEM certificate files
In addition to JKS keystores and trustores, from version 2.7.0 of Kafka,
certificates in PEM format can be used when enabling SSL.

For PEM files to be used KAFKA_CFG_TLS_TYPE must be set to PEM and
certificates must be named:
* kafka.keystore.pem
* kafka.truststore.pem

The keystore must contain both the public and private keys in PKCS#8
format, concatenated.

The private key can be encrypted. Set the password with the
KAFKA_CERTIFICATE_PASSWORD environment variable.