Add support for PEM certificate files
In addition to JKS keystores and trustores, from version 2.7.0 of Kafka,
certificates in PEM format can be used when enabling SSL.

For the PEM files to be recognized the following names must be used:
kafka.pem: public certificate (equivalent to part of the JKS keystore)
kafka.key: private key (equivalent to part of the JKS keystore)
ca-cert: public CA certificate (equivalent to the JKS truststore)

The private key can be encrypted. Set the password with the
KAFKA_CERTIFICATE_PASSWORD environment variable.

The keys and certificates will be stored as a multi-line string in the
properties files. kafka_common_conf_set will use a perl one-line
script to replace multi-line values if needed.